BUG#: 4451 TITLE: Compile failure building CIMOperationResponseEncode.cpp
DESCRIPTION: Changed type from Uint32 to Uint64.

+++ src/Pegasus/Server/CIMOperationResponseEncoder.cpp
@@ -160,7 +160,7 @@ CIMOperationResponseEncoder::sendResponse(CIMResponseMessage* response,
 																 HttpMethod httpMethod,
 																 const ContentLanguages &httpContentLanguages, 
 																 const Buffer& body,
-																 Uint32 serverResponseTime,
+																 Uint64 serverResponseTime,
 																 Boolean isFirst,
 																 Boolean isLast);
